Rewrite the ClangSharpPInvokeGenerator to allow better reuse and to be more explicit (#34)
* Performing some cleanup of the OOP methods, primarily to fix debugging

* Adding a launchSettings.json so that F5 regenerates ClangSharp

* Rewriting ClangSharpPInvokeGenerator

* Regenerating ClangSharp and updating OOP extensions

* Removing the dependency on S.R.CompilerServices.Unsafe and fixing the launchSettings to be machine independent

+++ ClangSharp/Extensions/CXTranslationUnit.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Runtime.CompilerServices;
 
 namespace ClangSharp
 {
@@ -48,7 +49,7 @@ namespace ClangSharp
 
         public CXFile GetFile(string fileName) => clang.getFile(this, fileName);
 
-        public string GetFileContents(CXFile file, out ulong size) => clang.getFileContents(this, file, out size);
+        public string GetFileContents(CXFile file, out IntPtr size) => clang.getFileContents(this, file, out size);
 
         public void GetInclusions(CXInclusionVisitor visitor, CXClientData clientData) => clang.getInclusions(this, visitor, clientData);
